refactor(core): move preflight to mod and separate into util
