+++ core/src/lib.rs
@@ -1,26 +1,23 @@
-use crate::{
-    interfaces::{ProofRequest, RaikoError, RaikoResult},
-    preflight::preflight,
-    provider::BlockDataProvider,
-};
+use std::{collections::HashMap, hint::black_box};
+
 use alloy_primitives::Address;
 use alloy_rpc_types::EIP1186AccountProofResponse;
-use preflight::PreflightData;
-use raiko_lib::protocol_instance::ProtocolInstance;
-use raiko_lib::prover::Proof;
 use raiko_lib::{
     builder::{create_mem_db, RethBlockBuilder},
-    prover::ProofKey,
-};
-use raiko_lib::{
     consts::ChainSpec,
     input::{GuestInput, GuestOutput, TaikoProverData},
-    prover::{IdStore, IdWrite},
+    protocol_instance::ProtocolInstance,
+    prover::{IdStore, IdWrite, Proof, ProofKey},
 };
 use reth_primitives::Header;
 use serde_json::Value;
-use std::{collections::HashMap, hint::black_box};
 use tracing::{debug, error, info, warn};
+
+use crate::{
+    interfaces::{ProofRequest, RaikoError, RaikoResult},
+    preflight::{preflight, PreflightData},
+    provider::BlockDataProvider,
+};
 
 pub mod interfaces;
 pub mod preflight;
